pyproject.toml
[project]
name = "vanta-cli"
version = "0.1.0"
description = "CLI tool for the Vanta compliance API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.2",
    "httpx>=0.28",
    "typer>=0.15",
    "rich>=13.0",
]

[project.scripts]
vanta = "vanta_cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vanta_cli"]
